+++ e2e-tests/page-objects/paste-window/paste-window.ts
@@ -1,0 +1,7 @@
+import { BaseWindow } from "../base-window";
+
+export abstract class PasteWindow extends BaseWindow {
+  static async activate(): Promise<void> {
+    await this.baseActivate('http://tauri.localhost/paste-window');
+  }
+}