+++ release-notes.md
@@ -5,3 +5,5 @@
 ## Fixes
 
 * Improved icons preloading.
+* Improved database performance.
+* Fixed installed plugin card enabled/disabled indicator responsiveness.